Change color and remove asterisk
Since we know from testing with a modified direwolf (PR proposed) that
the asterisk comes through, this commit removes the star associated with
the coloring. Also modify the coloring. The red was too strong. Now the
received digi is light/cyan/bold and the not-received digis are
light/gray/bold.