Try an alternate codec for connected data if utf-8 fails
If decoding connected data as utf-8 fails, try a configurable alternate
codec, and if that fails, decode as utf-8 with replacement characters.
Document and explain the config option, but don't expose it through the
UI, since it's rather esoteric and likely to be very rarely used.

Thanks to @esev for the proposal.

Fixes #10, Resolves #11